explicit conversions and fix ZeroDivisionError in mosaic slider

+++ vini/SliceWindow.py
@@ -28,7 +28,7 @@ class SliceWindow(QtGui.QWidget):
         size = self.geometry()
         self.id = window_number
         # Place it in the center of the screen.
-        self.move((screen.width()-size.width())/2, (screen.height()-size.height())/2)
+        self.move(int((screen.width()-size.width())/2), int((screen.height()-size.height())/2))
 
         self.l = QtGui.QGridLayout()
         self.setLayout(self.l)
